// src/app/shared/models/peoducts.model.ts
export class Products {

    constructor(){
        this.imageExtension = 'png';

    }
    id: number;;
    name: string;
    imagePath: string;
    price: number;;
    categoryName: string;
    supplierName: string;
    categoryId: number;
    supplierId: number;
    order: number;
    imageExtension: string;
    imageBase64: string;
}